requirements.txt
black==19.10b0
flake8==3.8.3
PyYAML==5.3.1
requests==2.24.0
gitpython==3.1.11